Rename config to user
